feat(web+server): fix Alphapolis
- 服务端 points 计数改成 long
- Alphapolis 反爬需要用 tabFetch
  或者用户需要至少保持一个 Alphapolis Tab open）

+++ server/src/main/kotlin/infra/web/datasource/providers/Novelup.kt
@@ -75,7 +75,7 @@ class Novelup(
         val points = row("応援ポイント")
             .text()
             .filter { it.isDigit() }
-            .toInt()
+            .toLong()
 
         val totalCharacters = row("文字数")
             .text()
